Make HIDDEN fields carry real data; fix zoned/packed dec==0 type bug
HIDDEN fields never actually carried a value through the buffer, despite
real DDS "H" only meaning "not rendered," not "no data" — the classic
SFLRCDNBR HIDDEN pattern (read back which subfile row was selected) never
worked. Three layers all had to skip H consistently for this to go
unnoticed: dspf__extractFields/dspf__applyFields only advanced the
buffer pointer past H fields without reading/writing their bytes, and
_dspf.h's struct generation gave them no member at all. Fixed both here;
OpenRPG's codegen side (declaration + copy loops) is fixed in a
companion commit there.

Fixing this surfaced a second, unrelated, wider bug: cppFieldType()
special-cased zoned/packed fields with zero decimal places as C++ long,
while OpenRPG's codegen and this runtime's own buffer-walking code always
assume double for those types regardless of decimals. The mismatch
silently type-punned a double-encoded buffer slot as a 64-bit integer
wherever generated C++ read it directly (e.g. `SFLRCDNBR = ctlBuf_.SFLRCDNBR;`
producing 4607182418800017408.0 — 1.0's bit pattern reinterpreted as an
int64). Not HIDDEN-specific: it affected any whole-number zoned/packed
WORKSTN field. Never caught before because no existing test read such a
field's value back, only its presence or an indicator it set. Fixed by
always returning "double", matching the assumption everywhere else;
regenerated 5 golden _dspf.h files whose struct layout changed as a
result (test04, test07, TEST15 x2, TEST17).

test17 now actually reads SFLRCDNBR back after EXFMT (previously had to
avoid it, since it didn't work) and confirms it comes back as 1 —
verifying both fixes end to end.

+++ tests/expected/TEST15_CHANGE_BLANKS_fixed_dspf.h
@@ -7,7 +7,7 @@
 
 struct CHGBLKTEST_buf {
     char FLDX[6] = {};
-    long FLDQTY = 0;
+    double FLDQTY = 0.0;
 };
 
 #endif // TEST15_CHANGE_BLANKS_FIXED_DSPF_H
